reverted ga_resources to track github

+++ ga_resources/signals.py
@@ -1,6 +1,7 @@
+import datetime
+
 from django.db.models.signals import post_save, pre_delete, pre_save
 from django.conf import settings as s
-import datetime
 from django.utils.timezone import utc
 from . import tasks, dispatch, drivers
 from .models import DataResource, Style, RenderedLayer
@@ -23,16 +24,25 @@ def dataresource_post_save(sender, instance, *args, **kwargs):
 
 
 def purge_cache_on_delete(sender, instance, *args, **kwargs):
-    instance.resource.clear_cache()
-    if instance.resource_file and os.path.exists(os.path.join(s.MEDIA_ROOT, instance.resource_file.name)):
-        os.unlink(os.path.join(s.MEDIA_ROOT, instance.resource_file.name))
+    if sender is DataResource:
+        instance.resource.clear_cache()
+        if instance.resource_file and os.path.exists(os.path.join(s.MEDIA_ROOT, instance.resource_file.name)):
+            os.unlink(os.path.join(s.MEDIA_ROOT, instance.resource_file.name))
+    elif sender is Style:
+        for layer in instance.renderedlayer_set.all():
+            layer.resource.clear_cache()
+    elif sender is RenderedLayer:
+        layer.resource.clear_cache()
+
 
 def purge_resource_data(sender, instance, *args, **kwargs):
     if sender is DataResource:
         instance.driver_instance.clear_cache()
 
+
 def shave_tile_caches(sender, instance, bbox, *args, **kwargs):
     drivers.CacheManager.get().shave_caches(instance, bbox)
+
 
 def trim_tile_caches(sender, instance, *args, **kwargs):
     if sender is Style:
